feat(textures): move texture elements to separate files
