+++ textures/Circle.test.js
@@ -1,6 +1,6 @@
 import lng from '@lightningjs/core';
 import TestUtils from '../test/lightning-test-utils';
-import { Arrow, Circle } from '.';
+import Circle from './Circle';
 
 const createTexture = TestUtils.makeCreateComponent(lng.Component);
 
@@ -30,32 +30,5 @@ describe('Circle', () => {
     expect(circle.stroke).toBe(false);
     expect(circle.strokeColor).toBe('rgb(0,0,0)');
     expect(circle.strokeWidth).toBe(1);
-  });
-});
-
-describe('Arrow', () => {
-  let arrow, component, testRenderer;
-
-  beforeEach(() => {
-    [component, testRenderer] = createTexture({
-      texture: {
-        type: Arrow,
-        w: 10,
-        h: 10,
-        direction: 'right'
-      }
-    });
-    arrow = component.texture;
-  });
-
-  it('renders', () => {
-    const tree = testRenderer.toJSON(2);
-    expect(tree).toMatchSnapshot();
-  });
-
-  it('has props', () => {
-    expect(arrow.w).toBe(10);
-    expect(arrow.h).toBe(10);
-    expect(arrow.direction).toBe('right');
   });
 });
